# api.yaml
openapi: 3.0.3
info:
  title: Room Booking Service (Test Task)
  description: Сервис бронирования переговорок
  version: "1.0.0"
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoomIdPath:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Идентификатор переговорки
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - NOT_FOUND
                - ROOM_NOT_FOUND
                - SLOT_NOT_FOUND
                - SLOT_ALREADY_BOOKED
                - BOOKING_NOT_FOUND
                - FORBIDDEN
                - SCHEDULE_EXISTS
                - INTERNAL_ERROR
            message:
              type: string
      example:
        error:
          code: INVALID_REQUEST
          message: invalid request
    InternalErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INTERNAL_ERROR
            message:
              type: string
      example:
        error:
          code: INTERNAL_ERROR
          message: internal server error
    Token:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: JWT для заголовка Authorization Bearer
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user]
        createdAt:
          type: string
          format: date-time
          nullable: true
...
    Schedule:
      type: object
      required: [roomId, daysOfWeek, startTime, endTime]
      properties:
        id:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        daysOfWeek:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 7
          description: "1=Пн .. 7=Вс, дни доступности. Каждое значение должно быть в диапазоне 1–7."
        startTime:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: Время начала (HH:MM) в рамках дня. Интерпретируется как UTC.
        endTime:
          type: string
          pattern: "^([01]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: Время окончания (HH:MM) в рамках дня. Интерпретируется как UTC.
    Slot:
      type: object
      required: [id, roomId, start, end]
      properties:
        id:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
        start:
          type: string
          format: date-time
          description: Время начала слота в UTC (ISO 8601, например 2024-06-10T09:00:00Z)
        end:
          type: string
          format: date-time
          description: Время окончания слота в UTC (ISO 8601, например 2024-06-10T09:30:00Z)
...
paths:
...
  /register:
    post:
      tags: [Auth]
      summary: "Регистрация пользователя (ДОПОЛНИТЕЛЬНОЕ ЗАДАНИЕ — необязательно)"
      description: >
        Создаёт нового пользователя и возвращает его данные.
        Реализация этого эндпоинта является **дополнительным заданием**.
        Для авторизации в рамках обязательной части используйте `/dummyLogin`.
        Доступен без авторизации. В обычном режиме разрешена только роль `user`;
        самостоятельная регистрация `admin` доступна только при `TEST_TASK_MODE=true`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                role:
                  type: string
                  enum: [admin, user]
      responses:
        '201':
          description: Пользователь создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Неверный запрос или email уже занят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }
  /login:
    post:
      tags: [Auth]
      summary: "Авторизация по email и паролю (ДОПОЛНИТЕЛЬНОЕ ЗАДАНИЕ — необязательно)"
      description: >
        Авторизует пользователя по email и паролю, возвращает JWT.
        Реализация этого эндпоинта является **дополнительным заданием**.
        Для авторизации в рамках обязательной части используйте `/dummyLogin`.
        Доступен без авторизации.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }
  /dummyLogin:
    post:
      tags: [Auth]
      summary: Получить тестовый JWT по роли (ОБЯЗАТЕЛЬНО). Доступен без авторизации.
      description: >
        Выдаёт тестовый JWT для указанной роли (admin / user).
        Для каждой роли возвращается **фиксированный UUID** пользователя:
        один и тот же UUID для всех запросов с ролью admin и один и тот же UUID для роли user.
        Это обеспечивает стабильность при тестировании сценариев, требующих проверки владельца брони.
        JWT должен содержать как минимум `user_id` (UUID) и `role`.
        Этот эндпоинт является **обязательным** и должен быть реализован в любом случае,
        независимо от реализации `/register` и `/login`.
        Принимает только допустимые значения роли: `admin` или `user`; иные значения — 400.
        Маршрут регистрируется только при `TEST_TASK_MODE=true` и отсутствует в production mode.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, user]
      responses:
        '200':
          description: Тестовый токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Неверный запрос (недопустимое значение роли)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }

  # ──────────────────────────────────────────────────────────────
  # Rooms
  # ──────────────────────────────────────────────────────────────
...
  /rooms/{roomId}/schedule/create:
    post:
      tags: [Schedules]
      summary: Создать расписание переговорки (только admin, только один раз). Длительность слота 30 мин. После создания расписание изменить нельзя.
      description: >
        Доступно только роли admin.
        Расписание можно создать только один раз для каждой переговорки.
        Поле `daysOfWeek` должно содержать значения от 1 до 7 (1=Пн, 7=Вс).
        При передаче значений вне этого диапазона возвращается 400.
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        '201':
          description: Расписание сохранено
          content:
            application/json:
              schema:
                type: object
                properties:
                  schedule:
                    $ref: '#/components/schemas/Schedule'
        '400':
          description: Неверный запрос (в т.ч. недопустимые значения daysOfWeek)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Доступ запрещён (требуется роль admin)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Переговорка не найдена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Расписание для переговорки уже создано, изменение не допускается
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error:
                  code: SCHEDULE_EXISTS
                  message: schedule for this room already exists and cannot be changed
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }

  # ──────────────────────────────────────────────────────────────
  # Slots
  # ──────────────────────────────────────────────────────────────
  /rooms/{roomId}/slots/list:
    get:
      tags: [Slots]
      summary: Список доступных для бронирования слотов по переговорке и дате (admin и user). Наиболее нагруженный эндпоинт.
      description: >
        Возвращает слоты, не занятые активной бронью, для указанной переговорки на указанную дату.
        Все даты и время передаются и возвращаются в UTC.
        Параметр `date` является обязательным; при его отсутствии возвращается 400.
        Если у переговорки нет расписания — возвращается пустой список (переговорка считается всегда недоступной).
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RoomIdPath'
        - name: date
          in: query
          required: true
          description: "Дата в формате ISO 8601 (например: 2024-06-10). Обязательный параметр; при отсутствии — 400."
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Список доступных слотов (не занятых активной бронью). Пустой список, если у переговорки нет расписания или на эту дату нет слотов.
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/Slot'
        '400':
          description: Неверный запрос (отсутствует или некорректен параметр date)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Переговорка не найдена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }

  # ──────────────────────────────────────────────────────────────
  # Bookings
  # ──────────────────────────────────────────────────────────────
...